move testcase to relative file
